backend/requirements.txt
fastapi==0.121.2
pydantic==2.12.4
requests==2.32.5
python-dotenv==1.2.1
instaloader==4.15
TikTokApi==7.2.1
uvicorn[standard]==0.38.0
supabase==2.12.0
openai==1.59.5

